view and rollback

+++ WEB-INF/src/DeleteServlet.java
@@ -54,6 +54,7 @@ public class DeleteServlet extends HttpServlet {
 		Connection conn = null;
 		Statement stmt = null;
 		Statement stmt_hla = null;
+		Statement stmt_user = null;
 		try {
 			Class.forName("org.postgresql.Driver");
 			conn = DriverManager.getConnection("jdbc:postgresql://" + _hostname + ":5432/" + _dbname, _username,
@@ -66,25 +67,37 @@ public class DeleteServlet extends HttpServlet {
 			stmt.executeUpdate("DELETE FROM address WHERE user_id ='" + deleteuser_data + "'");
 			stmt.executeUpdate("DELETE FROM contact WHERE user_id ='" + deleteuser_data + "'");
 			stmt.executeUpdate("DELETE FROM login WHERE user_id ='" + deleteuser_data + "'");
-			ResultSet rs = stmt.executeQuery("SELECT * FROM register WHERE user_id ='" + deleteuser_data + "'");
-			rs.next();
-			int hla_id = rs.getInt("hla_id");
-			stmt.executeUpdate("DELETE FROM register WHERE user_id= '" + deleteuser_data + "'");
-			stmt.executeUpdate("DELETE FROM user_data WHERE user_id= '" + deleteuser_data + "'");
-			rs.close();
-			stmt_hla = conn.createStatement();
-			ResultSet rs_hla = stmt_hla.executeQuery("SELECT * FROM register WHERE hla_id ='" + hla_id + "'");
-			String same_user = "";
-			if (rs_hla.next()) {
-				same_user = rs_hla.getString("user_id");
+			stmt_user = conn.createStatement();
+			ResultSet rs_user = stmt_user
+					.executeQuery("select count(*) from user_data where user_id = '" + deleteuser_data + "'");
+			rs_user.next();
+			int count = rs_user.getInt("count");
+			if (count <= 0) {
+				conn.rollback();
+				out.println("<h3>該当するユーザは存在しません。</h3><br/>");
+			} else {
+				
+				out.println("<h3>以下のユーザを削除しました。</h3><br/>");
+				out.println("ユーザID: " + deleteuser_data + "<br/>");
+
+				ResultSet rs = stmt.executeQuery("SELECT * FROM register WHERE user_id ='" + deleteuser_data + "'");
+				rs.next();
+				int hla_id = rs.getInt("hla_id");
+				stmt.executeUpdate("DELETE FROM register WHERE user_id= '" + deleteuser_data + "'");
+				stmt.executeUpdate("DELETE FROM user_data WHERE user_id= '" + deleteuser_data + "'");
+				rs.close();
+				stmt_hla = conn.createStatement();
+				ResultSet rs_hla = stmt_hla.executeQuery("SELECT * FROM register WHERE hla_id ='" + hla_id + "'");
+				String same_user = "";
+				if (rs_hla.next()) {
+					same_user = rs_hla.getString("user_id");
+				}
+				if (same_user == "") {
+					stmt_hla.executeUpdate("DELETE FROM hla WHERE hla_id = '" + hla_id + "'");
+				}
+				conn.commit();
+				rs_hla.close();
 			}
-			if (same_user == "") {
-				stmt_hla.executeUpdate("DELETE FROM hla WHERE hla_id = '" + hla_id + "'");
-			}
-			rs_hla.close();
-			conn.commit();
-			out.println("<h3>以下のユーザを削除しました。</h3><br/>");
-			out.println("ユーザID: " + deleteuser_data + "<br/>");
 
 		} catch (Exception e) {
 			e.printStackTrace();
